Got error-free input smoothing running!
The input smoothing now outputs debugging messages indicating that it is running regularly.
There was an issue with the Bluetooth permissions inside Android which I resolved by copying the iBeacon package "Plugins" folder into the project root directory, causing Android to use the iBeacon AndroidManifest that includes Bluetooth permissions.
The input smoothing is running and checking, but no beacons are appearing in the input smoother at this time. Further testing is required.